refactor : filter에게 재발행 위임, 만료되지 않은 경우 재발행 api 호출 validation 처리
