+++ src/main/java/com/swef/cookcode/common/util/Util.java
@@ -1,13 +1,18 @@
 package com.swef.cookcode.common.util;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.swef.cookcode.common.ErrorCode;
 import com.swef.cookcode.common.UrlResponse;
 import com.swef.cookcode.common.error.exception.InvalidRequestException;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -16,6 +21,8 @@ import org.springframework.web.multipart.MultipartFile;
 public class Util {
 
     private final S3Util s3Util;
+
+    private final ObjectMapper objectMapper;
 
     public static <T> void validateDuplication(List<T> list1, List<T> list2) {
         Set<T> mergedSets = new HashSet<>() {{
@@ -44,6 +51,19 @@ public class Util {
             hasNext = true;
         }
         return hasNext;
+    }
+
+    public void setResponse(int status, HttpServletResponse response, Object responseBody) {
+        response.setStatus(status);
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        try {
+            String json = objectMapper.writeValueAsString(responseBody);
+            PrintWriter writer = response.getWriter();
+            writer.write(json);
+            writer.flush();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 
 }
