+++ src/main/java/com/swef/cookcode/user/controller/AccountController.java
@@ -247,10 +247,4 @@ public class AccountController {
         return ResponseEntity.ok(apiResponse);
     }
 
-//    @PostMapping("/token/reissue")
-//    public ResponseEntity<SignInResponse> reissueToken(@RequestParam(value = "refreshToken") String refreshToken) {
-//
-//        return null;
-//    }
-
 }
